+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Added JSON output to `6d6info` via the `--json` flag.
+- Added JSON progress output to `6d6copy` via the `--json-progress` flag.
 
 ## [1.5.1] - 2019-03-11
 ### Added